+++ src/util/UpdateCommonProps.ts
@@ -1,19 +1,25 @@
 import { modifiedDateTime } from "./generic";
 
+import { QlikRepositoryClient } from "../main";
+
 import { ICustomPropertyObject } from "../types/interfaces";
-
-import { ICustomPropertyCondensed } from "../CustomProperty";
+import {
+  ICustomPropertyCondensed,
+  IClassCustomProperty,
+  CustomProperty,
+} from "../CustomProperty";
 import { IAppUpdate } from "../App";
-import { ITagCondensed } from "../Tag";
-
+import { ITagCondensed, Tag, IClassTag } from "../Tag";
 import { ITaskCreate } from "../Task.interface";
-
-import { IUserUpdate } from "../User";
-import { IStreamUpdate } from "../Stream";
+import { IUserUpdate, IClassUser, User } from "../User";
+import { IStreamUpdate, Stream, IClassStream } from "../Stream";
 import { ISystemRuleCreate } from "../SystemRule.interface";
-
 export class UpdateCommonProperties {
-  private qlikUtil: any;
+  private qlikUtil: QlikRepositoryClient;
+  private customProperty: IClassCustomProperty;
+  private stream: IClassStream;
+  private tag: IClassTag;
+  private user: IClassUser;
   private arg:
     | IUserUpdate
     | IAppUpdate
@@ -34,6 +40,10 @@ export class UpdateCommonProperties {
     this.qlikUtil = qlikUtil;
     this.obj = obj;
     this.arg = arg;
+    this.customProperty = new CustomProperty(this.qlikUtil);
+    this.stream = new Stream(this.qlikUtil);
+    this.tag = new Tag(this.qlikUtil);
+    this.user = new User(this.qlikUtil);
   }
 
   async updateCustomProperties() {
@@ -43,8 +53,8 @@ export class UpdateCommonProperties {
       this.obj.customProperties = await Promise.all<ICustomPropertyObject>(
         this.arg.customProperties.map(async (customProperty) => {
           let [cpName, cpValue] = customProperty.split("=");
-          return await this.qlikUtil
-            .customPropertyGetFilter(`name eq '${cpName}'`)
+          return await this.customProperty
+            .getFilter(`name eq '${cpName}'`)
             .then((cpData) => {
               return {
                 definition: cpData[0] as ICustomPropertyCondensed,
@@ -61,8 +71,8 @@ export class UpdateCommonProperties {
     if (this.arg.tags && this.arg.tags.length > 0) {
       this.obj.tags = await Promise.all<ITagCondensed>(
         this.arg.tags.map(async (tag) => {
-          return await this.qlikUtil
-            .tagGetFilter(`name eq '${tag}'`)
+          return await this.tag
+            .getFilter(`name eq '${tag}'`)
             .then((tagsData) => tagsData[0] as ITagCondensed);
         })
       );
@@ -77,8 +87,8 @@ export class UpdateCommonProperties {
     if ((this.arg as IAppUpdate).owner) {
       let [userDirectory, userId] = (this.arg as IAppUpdate).owner.split("\\");
 
-      this.obj.owner = await this.qlikUtil
-        .userGetFilter(
+      this.obj.owner = await this.user
+        .getFilter(
           `userId  eq '${userId}' and userDirectory eq '${userDirectory}'`
         )
         .then((u) => u[0]);
@@ -87,8 +97,8 @@ export class UpdateCommonProperties {
 
   async updateAppStream() {
     if ((this.arg as IAppUpdate).stream) {
-      this.obj.stream = await this.qlikUtil
-        .streamGetFilter(`name eq '${(this.arg as IAppUpdate).stream}'`)
+      this.obj.stream = await this.stream
+        .getFilter(`name eq '${(this.arg as IAppUpdate).stream}'`)
         .then((streams) => streams[0]);
     }
   }